move: Footer 폴더 위치 pages -> components 폴더 하위로 이동
