+++ src/components/common/index.tsx
@@ -1,1 +1,2 @@
 export { default as Layout } from './Layout';
+export { default as Footer } from './Footer';
